+++ CHANGELOG.rst
@@ -66,6 +66,8 @@ Added
 * Creation of the ehv/hv grid model with osmTGmod, see 
   `issue #4 <https://github.com/openego/eGon-data/issues/4>`_ and 
   `PR #164 <https://github.com/openego/eGon-data/pull/164>`_  
+* Distribute electrical demands of households to zensus cells
+  `#181 <https://github.com/openego/eGon-data/issues/181>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159